[steps] move temporary files utils to root
